third_party/WebKit/LayoutTests/http/tests/credentialmanager/credentialscontainer-cross-context.html

<!doctype html>
<title>Credential Manager: Call get() across browsing contexts.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="/gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="/gen/third_party/WebKit/public/platform/modules/credentialmanager/credential_manager.mojom.js"></script>
<script src="/gen/third_party/WebKit/public/platform/modules/webauth/authenticator.mojom.js"></script>
<script src="resources/credential-helpers.js"></script>
<body>
<script>

var TEST_TOP_LEVEL_CREDENTIAL_ID = "topLevelCredentialId";
var TEST_NESTED_CREDENTIAL_ID = "nestedCredentialId";

mockCredentialManager.setResponse(TEST_TOP_LEVEL_CREDENTIAL_ID, "p", "", "");

var frame = document.createElement("iframe");
frame.src = "./resources/nested_credential_manager_client.html";
frame.addEventListener("load", _ => {

  promise_test(_ => {
    holder = navigator.credentials;
    receiver = navigator.credentials;
    return holder.get.call(receiver, {}).then(r => {
      assert_equals(r.id, TEST_TOP_LEVEL_CREDENTIAL_ID);
    });
  }, "top.navigator.credentials.get.call(top)");
  
  promise_test(_ => {
    holder = navigator.credentials;
    receiver = frame.contentWindow.navigator.credentials;
    return holder.get.call(receiver, {}).then(r => {
      assert_equals(r.id, TEST_TOP_LEVEL_CREDENTIAL_ID);
    });
  }, "top.navigator.credentials.get.call(nested)");
  
  promise_test(_ => {
    holder = frame.contentWindow.navigator.credentials;
    receiver = navigator.credentials;
    return holder.get.call(receiver, {}).then(r => {
      assert_equals(r.id, TEST_NESTED_CREDENTIAL_ID);
    });
  }, "nested.navigator.credentials.get.call(top)");
  
  promise_test(_ => {
    holder = frame.contentWindow.navigator.credentials;
    receiver = frame.contentWindow.navigator.credentials;
    return holder.get.call(receiver, {}).then(r => {
      assert_equals(r.id, TEST_NESTED_CREDENTIAL_ID);
    });
  }, "nested.navigator.credentials.get.call(nested)");

});
document.body.append(frame);
</script>
